# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Melogold API",
    "version": "1",
    "description": "Sync server of Melogold: favorites, library, playlists, history and playback across a user's devices. The normative contract is docs/API.md; clients branch only on `code` of ErrorResponse (error-codes.json). There is no version in the URL: compatibility comes from `apiVersion`/`minApiVersion` and `features.sync.protocol` of `GET /server/info` (API §1.1).",
    "license": {
      "name": "CC0-1.0",
      "url": "https://creativecommons.org/publicdomain/zero/1.0/"
    }
  },
  "servers": [
    {
      "url": "https://api.melogold.app",
      "description": "The official server. Self-hosted servers serve the same API under their own base URL."
    }
  ],
...
  "paths": {
...
    "/sync/summary": {
      "get": {
        "operationId": "getSyncSummary",
        "summary": "Counts of the server library for the merge dialog",
        "tags": [
          "sync"
        ],
        "description": "Rate limit: 10/min per user (API §1.10).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "in": "header",
            "name": "X-Sync-Protocol",
            "required": true,
            "description": "Sync protocol of the client, within [features.sync.minProtocol, features.sync.protocol] (API §1.2). Missing or not an integer → 400 invalid_request; outside → 409 protocol_unsupported."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncSummary"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request` (issues)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`, `access_token_invalid`, `access_token_expired`, `session_revoked`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "`protocol_unsupported` (minProtocol, maxProtocol)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal_error`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`server_busy` (retryAfterSeconds), `unavailable` (retryAfterSeconds?)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/sync/merge-plan": {
      "post": {
        "operationId": "planMerge",
        "summary": "Plan how local playlists merge with the server ones",
        "tags": [
          "sync"
        ],
        "description": "A pure function: nothing is written. `plan` follows the order of the request (API §4.7).\n\nRate limit: 10/min per user (API §1.10).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergePlanRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "in": "header",
            "name": "X-Sync-Protocol",
            "required": true,
            "description": "Sync protocol of the client, within [features.sync.minProtocol, features.sync.protocol] (API §1.2). Missing or not an integer → 400 invalid_request; outside → 409 protocol_unsupported."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MergePlanResponse"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request` (issues), `invalid_json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`, `access_token_invalid`, `access_token_expired`, `session_revoked`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "`protocol_unsupported` (minProtocol, maxProtocol)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "`payload_too_large`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "`unsupported_media_type`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal_error`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`server_busy` (retryAfterSeconds), `unavailable` (retryAfterSeconds?)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/sync": {
      "post": {
        "operationId": "sync",
        "summary": "Send ops and receive the changes since the cursor",
        "tags": [
          "sync"
        ],
        "description": "Ops are idempotent: a repeated op answers `replayed: true`. Only `opId`, `kind`, `at` and `base` of each op are validated here; a bad field of one op defers that op, never the batch. Work budget: Σ(videoIds + entries + tracks) ≤ 20 000, else 413 (API §1.9, §4.8).\n\nRate limit: 120/min per user (API §1.10).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^-?[0-9]+$"
            },
            "in": "header",
            "name": "X-Sync-Protocol",
            "required": true,
            "description": "Sync protocol of the client, within [features.sync.minProtocol, features.sync.protocol] (API §1.2). Missing or not an integer → 400 invalid_request; outside → 409 protocol_unsupported."
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponse"
                }
              }
            }
          },
          "400": {
            "description": "`invalid_request` (issues), `invalid_json`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized`, `access_token_invalid`, `access_token_expired`, `session_revoked`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "`protocol_unsupported` (minProtocol, maxProtocol)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "410": {
            "description": "`cursor_invalid`, `cursor_expired` (floorCursor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "`payload_too_large`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "`unsupported_media_type`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal_error`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "`server_busy` (retryAfterSeconds), `unavailable` (retryAfterSeconds?), `storage_full` (retryAfterSeconds)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token of TokenPair (API §1.7)."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "error": {
            "type": "string",
            "description": "Same text as `message` (Clementine shape)."
          },
          "message": {
            "type": "string",
            "description": "English text for logs; not for UI or logic. 5xx always carry the generic text of their code."
          },
          "code": {
            "type": "string",
            "description": "The only field clients branch on (API §2.2, `error-codes.json`). Clients must accept unknown codes."
          },
          "retryAfterSeconds": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            }
          },
          "minLength": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "maxLength": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "deviceLimit": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "deviceCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "minProtocol": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "maxProtocol": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "floorCursor": {
            "type": "string",
            "description": "Opaque cursor (API §1.6)."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "code"
        ],
        "description": "Every refusal (API §2.1). Details appear only with their codes."
      },
...
      "MergePlanEntry": {
        "type": "object",
        "properties": {
          "localKey": {
            "type": "string"
          },
          "action": {
            "type": "string",
            "description": "Known values: merge, deleted, create. Clients must accept unknown values (API §1.3)."
          },
          "playlistId": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "description": "UUID, lowercase."
          },
          "serverName": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "localKey",
          "action",
          "playlistId",
          "serverName"
        ]
      },
      "MergePlanInput": {
        "type": "object",
        "properties": {
          "localKey": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "description": "Unique within the request."
          },
          "syncId": {
            "nullable": true,
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "description": "UUID, lowercase only (API §1.6)."
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "browseId": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "localKey",
          "name"
        ]
      },
      "MergePlanRequest": {
        "type": "object",
        "properties": {
          "playlists": {
            "maxItems": 5000,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MergePlanInput"
            }
          }
        },
        "required": [
          "playlists"
        ]
      },
      "MergePlanResponse": {
        "type": "object",
        "properties": {
          "plan": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MergePlanEntry"
            },
            "description": "In the order of the request."
          }
        },
        "required": [
          "plan"
        ]
      },
      "OpResult": {
        "type": "object",
        "properties": {
          "opId": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "description": "UUID, lowercase."
          },
          "status": {
            "type": "string",
            "description": "Known values: applied, superseded, redirected, rejected, deferred. Clients must accept unknown values (API §1.3)."
          },
          "code": {
            "nullable": true,
            "type": "string",
            "description": "API §2.3. Known values: playlist_deleted, invalid_video_id, unknown_kind, invalid_payload, quota_exceeded, playlist_not_found, op_rate_limited. Clients must accept unknown values (API §1.3)."
          },
          "seq": {
            "nullable": true,
            "description": "For debugging.",
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "playlistId": {
            "nullable": true,
            "description": "Only `redirected`: the recovery playlist.",
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "retryAfterSeconds": {
            "nullable": true,
            "description": "Only `op_rate_limited`.",
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "replayed": {
            "type": "boolean"
          }
        },
        "required": [
          "opId",
          "status",
          "code",
          "seq",
          "playlistId",
          "retryAfterSeconds",
          "replayed"
        ]
      },
...
      "SyncInclude": {
        "type": "object",
        "properties": {
          "likes": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{11}$",
              "description": "Any YouTube video id (API §1.6)."
            }
          },
          "playlists": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "description": "UUID, lowercase only (API §1.6)."
            }
          },
          "bookmarks": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookmarkKey"
            }
          },
          "playStats": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{11}$",
              "description": "Any YouTube video id (API §1.6)."
            }
          }
        },
        "description": "Return the current rows of these keys (at most 1000 in total)."
      },
...
      "SyncOp": {
        "type": "object",
        "properties": {
          "opId": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "description": "UUID v4 made by the client; `play.add`: the eventId."
          },
          "kind": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "description": "One of features.sync.kinds: like.set, bookmark.set, playlist.create, playlist.update, playlist.delete, playlist.items.add, playlist.item.remove, playlist.item.move, playlist.items.replace, playlist.import, play.add, play.baseline, history.clear, history.forget. An unknown kind → deferred unknown_kind."
          },
          "at": {
            "format": "date-time",
            "pattern": "^(\\d{4})-(\\d{2})-(\\d{2})T(\\d{2}):(\\d{2}):(\\d{2})(?:\\.(\\d{1,9}))?Z$",
            "description": "When the user acted (client clock with clockOffset); `play.add`: = playedAt.",
            "type": "string"
          },
          "base": {
            "description": "The client's cursor when the user acted; used to resolve conflicts (DESIGN §3.4).",
            "nullable": true,
            "type": "string",
            "maxLength": 64
          },
          "videoId": {
            "nullable": true,
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{11}$",
            "description": "VideoId; an invalid one → `rejected invalid_video_id`."
          },
          "videoIds": {
            "nullable": true,
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{11}$",
              "description": "VideoId; an invalid one → `rejected invalid_video_id`."
            }
          },
          "after": {
            "nullable": true,
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{11}$",
            "description": "Anchor: insert right after this videoId."
          },
          "before": {
            "nullable": true,
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{11}$",
            "description": "Anchor: insert right before this videoId."
          },
          "liked": {
            "nullable": true,
            "type": "boolean"
          },
          "likedAt": {
            "nullable": true,
            "format": "date-time",
            "pattern": "^(\\d{4})-(\\d{2})-(\\d{2})T(\\d{2}):(\\d{2}):(\\d{2})(?:\\.(\\d{1,9}))?Z$",
            "description": "UTC time with `Z`, 0–9 fraction digits (truncated to ms), in [2000-01-01, 2100-01-01) (API §1.5).",
            "type": "string"
          },
          "type": {
            "nullable": true,
            "type": "string",
            "description": "Bookmark type: album | artist."
          },
          "browseId": {
            "nullable": true,
            "type": "string",
            "maxLength": 64
          },
          "bookmarked": {
            "nullable": true,
            "type": "boolean"
          },
          "bookmarkedAt": {
            "nullable": true,
            "format": "date-time",
            "pattern": "^(\\d{4})-(\\d{2})-(\\d{2})T(\\d{2}):(\\d{2}):(\\d{2})(?:\\.(\\d{1,9}))?Z$",
            "description": "UTC time with `Z`, 0–9 fraction digits (truncated to ms), in [2000-01-01, 2100-01-01) (API §1.5).",
            "type": "string"
          },
          "title": {
            "nullable": true,
            "type": "string",
            "maxLength": 500
          },
          "subtitle": {
            "nullable": true,
            "type": "string",
            "maxLength": 500
          },
          "thumbnailUrl": {
            "nullable": true,
            "type": "string",
            "maxLength": 2048
          },
          "year": {
            "nullable": true,
            "type": "string",
            "maxLength": 16
          },
          "playlistId": {
            "nullable": true,
            "type": "string",
            "description": "Uuid of the playlist."
          },
          "name": {
            "nullable": true,
            "type": "string",
            "description": "Playlist name: trimmed, cut to 200; empty → «Без названия» / «Untitled»."
          },
          "playedAt": {
            "nullable": true,
            "format": "date-time",
            "pattern": "^(\\d{4})-(\\d{2})-(\\d{2})T(\\d{2}):(\\d{2}):(\\d{2})(?:\\.(\\d{1,9}))?Z$",
            "description": "UTC time with `Z`, 0–9 fraction digits (truncated to ms), in [2000-01-01, 2100-01-01) (API §1.5).",
            "type": "string"
          },
          "playTimeMs": {
            "nullable": true,
            "type": "integer",
            "minimum": 1,
            "maximum": 86400000,
            "format": "int32"
          },
          "history": {
            "nullable": true,
            "type": "boolean"
          },
          "playtime": {
            "nullable": true,
            "type": "boolean"
          },
          "mode": {
            "nullable": true,
            "type": "string",
            "description": "play.baseline: add | atLeast."
          },
          "entries": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BaselineEntry"
            },
            "minItems": 1,
            "maxItems": 500
          },
          "eventsBefore": {
            "nullable": true,
            "format": "date-time",
            "pattern": "^(\\d{4})-(\\d{2})-(\\d{2})T(\\d{2}):(\\d{2}):(\\d{2})(?:\\.(\\d{1,9}))?Z$",
            "description": "Inclusive watermark.",
            "type": "string"
          },
          "resetTotal": {
            "nullable": true,
            "type": "boolean"
          },
          "tracks": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrackInput"
            },
            "description": "Metadata of the videoIds the op mentions."
          }
        },
        "required": [
          "opId",
          "kind",
          "at"
        ],
        "description": "Flat op (API §4.8). The route checks only opId, kind, at, base and unique opIds; other fields are checked per kind and answer in OpResult."
      },
      "SyncRequest": {
        "type": "object",
        "properties": {
          "cursor": {
            "type": "string",
            "pattern": "^(?:[0-9a-f]{8}\\.[0-9]{1,16}\\.[0-9]{1,16})?$",
            "description": "`\"\"` (both streams from zero) or an opaque cursor from a previous response (API §1.6)."
          },
          "limit": {
            "nullable": true,
            "type": "integer",
            "minimum": 1,
            "maximum": 2000,
            "format": "int32",
            "description": "Rows per page, default 500 (the first sync: 2000)."
          },
          "streams": {
            "nullable": true,
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "library",
                "history"
              ]
            },
            "uniqueItems": true,
            "description": "Non-empty subset; default: both."
          },
          "ops": {
            "nullable": true,
            "maxItems": 500,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SyncOp"
            }
          },
          "include": {
            "type": "object",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/SyncInclude"
              }
            ]
          }
        },
        "required": [
          "cursor"
        ]
      },
      "SyncResponse": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OpResult"
            },
            "description": "In the order of ops."
          },
          "cursor": {
            "type": "string",
            "description": "Opaque cursor (API §1.6)."
          },
          "hasMore": {
            "type": "boolean"
          },
          "serverTime": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$",
            "description": "UTC, always `YYYY-MM-DDTHH:mm:ss.sssZ` (API §1.5)."
          },
          "tracks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrackDto"
            },
            "description": "Ordered by videoId."
          },
          "playlists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaylistRow"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaylistItemRow"
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LikeRow"
            }
          },
          "bookmarks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookmarkRow"
            }
          },
          "plays": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayRow"
            }
          },
          "playStats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayStatRow"
            }
          },
          "playForgets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayForgetRow"
            }
          }
        },
        "required": [
          "results",
          "cursor",
          "hasMore",
          "serverTime",
          "tracks",
          "playlists",
          "items",
          "likes",
          "bookmarks",
          "plays",
          "playStats",
          "playForgets"
        ],
        "description": "Every row key appears once, as its full current image. Apply: tracks → playlists (by createdAt) → items → likes → bookmarks → playStats → plays → playForgets."
      },
      "SyncSummary": {
        "type": "object",
        "properties": {
          "cursor": {
            "type": "string",
            "description": "Opaque cursor (API §1.6)."
          },
          "serverTime": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$",
            "description": "UTC, always `YYYY-MM-DDTHH:mm:ss.sssZ` (API §1.5)."
          },
          "counts": {
            "$ref": "#/components/schemas/SyncSummaryCounts"
          }
        },
        "required": [
          "cursor",
          "serverTime",
          "counts"
        ],
        "description": "Numbers for the merge dialog (API §4.7)."
      },
      "SyncSummaryCounts": {
        "type": "object",
        "properties": {
          "likes": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "albums": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "artists": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "playlists": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "items": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "plays": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          },
          "playedTracks": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "format": "int64"
          }
        },
        "required": [
          "likes",
          "albums",
          "artists",
          "playlists",
          "items",
          "plays",
          "playedTracks"
        ]
      },
...
      "ValidationIssue": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "description": "Dot path into the request, e.g. \"ops.3.opId\"."
          },
          "code": {
            "type": "string",
            "description": "zod issue code, e.g. invalid_type, too_big, invalid_format."
          }
        },
        "required": [
          "path",
          "code"
        ]
      }
    }
  }
}